remove map duplicate code
