+++ nginx-controller/nginx/ingress.go
@@ -1,9 +1,6 @@
 package nginx
 
 import (
-	"fmt"
-	"strings"
-
 	api_v1 "k8s.io/client-go/pkg/api/v1"
 	extensions "k8s.io/client-go/pkg/apis/extensions/v1beta1"
 )
@@ -16,22 +13,4 @@ type IngressEx struct {
 	JWTKey        *api_v1.Secret
 	Endpoints     map[string][]string
 	EndpointsInfo map[string][]EndpointInfo
-}
-
-// TODO
-type EndpointInfo struct {
-	Address string
-	Info    map[string]interface{}
-}
-
-type Labels map[string]string
-
-func (labels Labels) String() string {
-	result := []string{}
-
-	for key, value := range labels {
-		result = append(result, fmt.Sprintf("%s:%s", key, value))
-	}
-
-	return strings.Join(result, ",")
 }
